+++ bookduck/src/main/java/com/mmc/bookduck/domain/archive/service/ExcerptService.java
@@ -7,7 +7,7 @@ import com.mmc.bookduck.domain.archive.repository.ExcerptRepository;
 import com.mmc.bookduck.domain.book.entity.UserBook;
 import com.mmc.bookduck.domain.book.service.UserBookService;
 import com.mmc.bookduck.domain.common.Visibility;
-import com.mmc.bookduck.domain.userhome.dto.common.ExcerptWithBookInfoUnitDto;
+import com.mmc.bookduck.domain.homecard.dto.common.ExcerptWithBookInfoUnitDto;
 import com.mmc.bookduck.domain.user.entity.User;
 import com.mmc.bookduck.domain.user.service.UserService;
 import com.mmc.bookduck.global.common.PaginatedResponseDto;
