+++ bookduck/src/main/java/com/mmc/bookduck/domain/homecard/dto/common/ExcerptWithBookInfoUnitDto.java
@@ -1,4 +1,4 @@
-package com.mmc.bookduck.domain.userhome.dto.common;
+package com.mmc.bookduck.domain.homecard.dto.common;
 
 import com.mmc.bookduck.domain.archive.entity.Excerpt;
 import com.mmc.bookduck.domain.common.Visibility;
