+++ bookduck/src/main/java/com/mmc/bookduck/domain/homecard/dto/common/HomeCardDto.java
@@ -1,4 +1,4 @@
-package com.mmc.bookduck.domain.userhome.dto.common;
+package com.mmc.bookduck.domain.homecard.dto.common;
 
 public sealed interface HomeCardDto permits ExcerptCardDto, OneLineCardDto, BookWithMemoCardDto, BookWithSongCardDto {}
