+++ bookduck/src/main/java/com/mmc/bookduck/domain/homecard/dto/common/HomeCardUpdateUnitDto.java
@@ -1,4 +1,4 @@
-package com.mmc.bookduck.domain.userhome.dto.common;
+package com.mmc.bookduck.domain.homecard.dto.common;
 
 public record HomeCardUpdateUnitDto(
         Long cardId,
