+++ bookduck/src/main/java/com/mmc/bookduck/domain/homecard/dto/request/ReadingSpaceUpdateRequestDto.java
@@ -1,0 +1,10 @@
+package com.mmc.bookduck.domain.homecard.dto.request;
+
+import com.mmc.bookduck.domain.homecard.dto.common.HomeCardUpdateUnitDto;
+
+import java.util.List;
+
+public record ReadingSpaceUpdateRequestDto(
+        List<HomeCardUpdateUnitDto> updatedCardList
+) {
+}
