+++ bookduck/src/main/java/com/mmc/bookduck/domain/homecard/dto/response/ReadingSpaceResponseDto.java
@@ -1,0 +1,10 @@
+package com.mmc.bookduck.domain.homecard.dto.response;
+
+import com.mmc.bookduck.domain.homecard.dto.common.HomeCardDto;
+
+import java.util.List;
+
+public record ReadingSpaceResponseDto(
+        List<HomeCardDto> cardList
+) {
+}
