+++ bookduck/src/main/java/com/mmc/bookduck/domain/homecard/entity/CardType.java
@@ -1,4 +1,4 @@
-package com.mmc.bookduck.domain.userhome.entity;
+package com.mmc.bookduck.domain.homecard.entity;
 
 public enum CardType {
     BOOK_WITH_MEMO,
