+++ bookduck/src/main/java/com/mmc/bookduck/domain/homecard/repository/HomeCardRepository.java
@@ -1,7 +1,7 @@
-package com.mmc.bookduck.domain.userhome.repository;
+package com.mmc.bookduck.domain.homecard.repository;
 
 import com.mmc.bookduck.domain.user.entity.User;
-import com.mmc.bookduck.domain.userhome.entity.HomeCard;
+import com.mmc.bookduck.domain.homecard.entity.HomeCard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
