--- bookduck/src/main/java/com/mmc/bookduck/domain/userhome/dto/request/ReadingSpaceUpdateRequestDto.java
@@ -1,10 +1,0 @@
-package com.mmc.bookduck.domain.userhome.dto.request;
-
-import com.mmc.bookduck.domain.userhome.dto.common.HomeCardUpdateUnitDto;
-
-import java.util.List;
-
-public record ReadingSpaceUpdateRequestDto(
-        List<HomeCardUpdateUnitDto> updatedCardList
-) {
-}
