--- bookduck/src/main/java/com/mmc/bookduck/domain/userhome/dto/response/ReadingSpaceResponseDto.java
@@ -1,10 +1,0 @@
-package com.mmc.bookduck.domain.userhome.dto.response;
-
-import com.mmc.bookduck.domain.userhome.dto.common.HomeCardDto;
-
-import java.util.List;
-
-public record ReadingSpaceResponseDto(
-        List<HomeCardDto> cardList
-) {
-}
